Fix minor leak, norme

+++ srcs/send_icmp_echo.c
@@ -1,3 +1,15 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   send_icmp_echo.c                                   :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jjaniec <jjaniec@student.42.fr>            +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2020/01/03 22:50:32 by jjaniec           #+#    #+#             */
+/*   Updated: 2020/01/03 23:00:11 by jjaniec          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include <ft_ping.h>
 
 extern t_ft_ping_info	*g_ft_ping_info;
@@ -103,7 +115,7 @@ ssize_t				send_icmp_echo(int s, struct sockaddr_in *dest)
 	else
 		dprintf(2, "Sendto() error\n");
 	g_ft_ping_info->wait_for_sigalrm = true;
-	if (g_ft_ping_info->pck_transmitted < g_ft_ping_info->pck_count)
+	if (g_ft_ping_info->pck_transmitted < (unsigned int)g_ft_ping_info->pck_count)
 		alarm(1);
 	return (r);
 }
